Add a base dataset class that loads arrow files
I'm hesitant to add mock data generators that would be required for unit testing this class. I suspect if we write 'composable' datasets and keep the actual data loading pretty isolated from any other logic (packing, shuffling, sharding) we can test the specific functionality of composing datasets separately. If we do add mock datasets, they won't cover the s3 path, which is the majority of what this class does.

I did test manually of course, and it would be pretty easy to add the code to generate mock datasets if we think it's necessary.

ghstack-source-id: 5fa9794fd1090c1f5fcd0c8e454a53262d568b54
Pull Request resolved: https://github.com/foundation-model-stack/foundation-model-stack/pull/124