Avoid double failure message
When the main action step of land or rebase fails, two steps are
executed to provide a comment on the PR with the link to the logs.

When a precondition fails, it posts an error message as comment to
the PR, which results into two comments added to the PR in this case.

To avoid this, further filter the execution of the error reporting
steps to only run when the main action step fails, and not when the
preconditions fail.

Signed-off-by: Andrea Frittoli <andrea.frittoligmail.com>

ghstack-source-id: aa03ff1b593b995c133eb18df3ecf765ef797570
Pull Request resolved: https://github.com/foundation-model-stack/foundation-model-stack/pull/157