Cache python dependencies in workflows
The fms-testing and mypy CI job install FMS dependencies and test
dependencies for every CI job, which is time and resource consuming.

Start caching the python virtualenv using GitHub caches. Introduce
requirement files with pinned dependencies which are then used
as keys to distinguish cache hits or miss.

Rather than caching the pip cache, we cache the entire virtualenv
as that contains wheels that might otherwise have to be build on
the fly for projects that don't provide one.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

[ghstack-poisoned]