Cache python dependencies in workflows
The fms-testing and mypy CI job install FMS dependencies and test
dependencies for every CI job, which is time and resource consuming.

Start caching the python virtualenv using GitHub caches. Introduce
requirement files with pinned dependencies which are then used
as keys to distinguish cache hits or miss.

Rather than caching the pip cache, we cache the entire virtualenv
as that further saves the installation time.

ghstack-source-id: fc0020964433f789ca43ae2a69fdaa7f31dd66d1
Pull Request resolved: https://github.com/foundation-model-stack/foundation-model-stack/pull/174

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>