Adapt land and rebase for GitHub workflow
Contributors to this repo may or may not adopt the GHStack workflow.
This extends the land and rebase chatops command to work with both
the GHStack workflow as well as the regular GitHub one.

The python script has been adapted to expose environment variables
to the GitHub actions, so that they have enough context to run both
the ghstack and the github workflows.

The GitHub workflow is implemented as much as possible using the gh
client calling into the GitHub API. This way the behaviour is
similar to that obtained by rebasing and merging via the GitHub UI.

Merging uses the "rebase" strategy only, without squash. PR authors
and reviewers are responsible for ensuring that the commits in a PR
are in good shape before the PR is landed.

Fixes: #166

ghstack-source-id: 5c4efdd8dc5f29f43ecccc641007b7f101d96def
Pull Request resolved: https://github.com/foundation-model-stack/foundation-model-stack/pull/167

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>