Add minimal torch.compile support to PagedTensor
I'm not very confident in this approach since references to _underlying_tensor
may not know about changes to PagedStorage during expansion etc
performed in other PagedTensors.

So-far the compile support added here doesn't handle any operations that
alter storage, so this is probably OK for now, but not sure how to
extend this to more complete support.

ghstack-source-id: e13cfd1b9bef58a0834e8645fc8ee802484dd7b2
Pull Request resolved: https://github.com/ibm-pytorch/foundation-model-stack/pull/254